+++ client/src/views/Register.vue
@@ -18,9 +18,17 @@
             <div class="textbox">
                 <input
                     type="password"
-                    id="register-email"
+                    id="register-password"
                     placeholder="Password"
                     v-model="register_password"
+                />
+            </div>
+            <div class="textbox">
+                <input
+                    type="password"
+                    id="register-password"
+                    placeholder="Confirm Password"
+                    v-model="register_password_confirm"
                 />
             </div>
             <button type="submit" class="btn register-btn">Sign Up</button>
@@ -29,9 +37,9 @@
             <router-link to="/login">
                 <span>Click Here to Login</span>
             </router-link>
-            <router-link to="/">
+            <!-- <router-link to="/">
                 <span>Click Here to Sign In with Google</span>
-            </router-link>
+            </router-link>-->
         </div>
     </div>
 </template>
@@ -44,35 +52,45 @@ export default {
         return {
             register_name: "",
             register_email: "",
-            register_password: ""
+            register_password: "",
+            register_password_confirm: ""
         };
     },
     methods: {
         register: function() {
-            Axios({
-                method: "post",
-                url: `${this.$store.state.rootUrl}/user/registration`,
-                data: {
-                    name: this.register_name,
-                    email: this.register_email,
-                    password: this.register_password
-                }
-            })
-                .then(result => {
-                    localStorage.setItem(
-                        "access_token",
-                        result.data.access_token
-                    );
-                    this.register_name = "";
-                    this.register_email = "";
-                    this.register_password = "";
-                    this.$store.state.isLogin = true;
-                    this.$router.push({ name: "Product" });
+            if (this.register_password == this.register_password_confirm) {
+                Axios({
+                    method: "post",
+                    url: `${this.$store.state.rootUrl}/user/registration`,
+                    data: {
+                        name: this.register_name,
+                        email: this.register_email,
+                        password: this.register_password
+                    }
                 })
-                .catch(err => {
-                    console.log(err.response);
-                    this.$store.dispatch("errorHandler", err.response);
+                    .then(result => {
+                        localStorage.setItem(
+                            "access_token",
+                            result.data.access_token
+                        );
+                        this.register_name = "";
+                        this.register_email = "";
+                        this.register_password = "";
+                        this.register_password_confirm = "";
+                        this.$store.state.isLogin = true;
+                        this.$router.push({ name: "Product" });
+                    })
+                    .catch(err => {
+                        console.log(err.response);
+                        this.$store.dispatch("errorHandler", err.response);
+                    });
+            } else {
+                this.$store.dispatch("errorHandler", {
+                    data: {
+                        message: "Confirmation password doesn't match"
+                    }
                 });
+            }
         }
     }
 };
